@@ -1,14 +1,15 @@
-const mongoose = require('mongoose')
-
-const cardSchema = new mongoose.Schema({
-    uid: { type: String, unique: true},
-    is_system: {type: Boolean,  default: false},
-    is_used: {type: Boolean,  default: false}
-})
-
-module.exports = mongoose.model('Card', cardSchema)
-
-
-
-
 const Sequelize = require('sequelize')
+
+module.exports.CardDbModel= ["cards", {
+    i_card: {
+        type: Sequelize.INTEGER,
+        autoIncrement: true,
+        primaryKey: true,
+        allowNull: false
+    },
+    uid: {
+        type: Sequelize.STRING,
+        allowNull: false
+    }
+}];
+
