@@ -5,26 +5,26 @@ import { MatTable } from '@angular/material/table';
 import { TestTableNgDataSource, TestTableNgItem } from './test-table-ng-datasource';
 
 @Component({
-  selector: 'app-test-table-ng',
-  templateUrl: './test-table-ng.component.html',
-  styleUrls: ['./test-table-ng.component.scss']
+    selector: 'app-test-table-ng',
+    templateUrl: './test-table-ng.component.html',
+    styleUrls: ['./test-table-ng.component.scss'],
 })
 export class TestTableNgComponent implements AfterViewInit, OnInit {
-  @ViewChild(MatPaginator, {static: false}) paginator: MatPaginator;
-  @ViewChild(MatSort, {static: false}) sort: MatSort;
-  @ViewChild(MatTable, {static: false}) table: MatTable<TestTableNgItem>;
-  dataSource: TestTableNgDataSource;
+    @ViewChild(MatPaginator, { static: false }) paginator: MatPaginator;
+    @ViewChild(MatSort, { static: false }) sort: MatSort;
+    @ViewChild(MatTable, { static: false }) table: MatTable<TestTableNgItem>;
+    dataSource: TestTableNgDataSource;
 
-  /** Columns displayed in the table. Columns IDs can be added, removed, or reordered. */
-  displayedColumns = ['id', 'name'];
+    /** Columns displayed in the table. Columns IDs can be added, removed, or reordered. */
+    displayedColumns = ['id', 'name'];
 
-  ngOnInit() {
-    this.dataSource = new TestTableNgDataSource();
-  }
+    ngOnInit() {
+        this.dataSource = new TestTableNgDataSource();
+    }
 
-  ngAfterViewInit() {
-    this.dataSource.sort = this.sort;
-    this.dataSource.paginator = this.paginator;
-    this.table.dataSource = this.dataSource;
-  }
+    ngAfterViewInit() {
+        this.dataSource.sort = this.sort;
+        this.dataSource.paginator = this.paginator;
+        this.table.dataSource = this.dataSource;
+    }
 }
